arogya/Aarogya-Sahayak/backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
beautifulsoup4
requests
pydantic
